:recycle: refactor: #9 이슈 각 모듈로 코드 이동
semo-api: semo app Controller, Service
semo-core: semo domain Entity, Repository, DTO
semo-batch: semo batch server
