+++ semo-api/src/main/java/sandbox/semo/application/SemoServiceApplication.java
@@ -3,9 +3,11 @@ package sandbox.semo.application;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EnableJpaAuditing
 @EnableJpaRepositories(basePackages = "sandbox.semo.domain")
 @EntityScan(basePackages = "sandbox.semo.domain")
 public class SemoServiceApplication {
