+++ semo-api/src/main/java/sandbox/semo/application/form/service/CompanyFormService.java
@@ -1,0 +1,8 @@
+package sandbox.semo.application.form.service;
+
+import sandbox.semo.domain.form.dto.request.CompanyFormRegister;
+
+public interface CompanyFormService {
+
+    void companyRegister(CompanyFormRegister companyFormRegister);
+}
