+++ semo-core/src/main/java/sandbox/semo/domain/form/repository/CompanyFormRepository.java
@@ -1,8 +1,7 @@
-package sandbox.semo.form.repository;
+package sandbox.semo.domain.form.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sandbox.semo.form.entity.CompanyForm;
+import sandbox.semo.domain.form.entity.CompanyForm;
 
 public interface CompanyFormRepository extends JpaRepository<CompanyForm, Long> {
-
 }
