+++ semo-core/src/main/java/sandbox/semo/domain/form/repository/MemberFormRepository.java
@@ -1,9 +1,7 @@
-package sandbox.semo.form.repository;
+package sandbox.semo.domain.form.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sandbox.semo.form.entity.MemberForm;
+import sandbox.semo.domain.form.entity.MemberForm;
 
 public interface MemberFormRepository extends JpaRepository<MemberForm, Long> {
-
-
 }
