--- src/main/java/sandbox/semo/form/service/CompanyFormService.java
@@ -1,8 +1,0 @@
-package sandbox.semo.form.service;
-
-import sandbox.semo.form.dto.CompanyFormRegister;
-
-public interface CompanyFormService {
-
-    void companyRegister(CompanyFormRegister companyFormRegister);
-}
